pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hebbs"
version = "0.1.2"
description = "Python SDK for HEBBS cognitive memory engine"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
keywords = ["hebbs", "memory", "ai", "agent", "cognitive", "grpc"]
authors = [
    {name = "HEBBS AI"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "grpcio>=1.60",
    "protobuf>=4.25",
]

[project.optional-dependencies]
demo = [
    "click>=8.1",
    "rich>=13.0",
    "google-genai>=1.0",
    "openai>=1.30",
    "anthropic>=0.25",
    "tomli>=2.0; python_version < '3.11'",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "grpcio-tools>=1.60",
    "ruff>=0.4",
    "mypy>=1.10",
    "mypy-protobuf>=3.6",
]

[project.scripts]
hebbs-demo = "demo.cli:main"

[project.urls]
Homepage = "https://hebbs.ai"
Repository = "https://github.com/hebbs-ai/hebbs-python"
Documentation = "https://docs.hebbs.ai"

[tool.hatch.build.targets.wheel]
packages = ["src/hebbs", "demo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "requires_server: requires a running HEBBS server",
    "requires_gemini: requires GEMINI_API_KEY environment variable",
    "requires_openai: requires OPENAI_API_KEY environment variable",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
